1. Правки операций CRUD

+++ src/main/java/service/ProjectService.java
@@ -20,7 +20,7 @@ public class ProjectService {
 
     //создание project в ProjectRepositoryImpl
     public void create(String id, String name, String description, String dateStart, String dateFinish, String userId){
-        Integer idChecked;
+        int idChecked;
         if(id == null) idChecked = projectRepository.getRepositorySize();
         else idChecked = Integer.parseInt(id);
 
@@ -51,6 +51,10 @@ public class ProjectService {
 
     //READ
     public Project findByName(String projectName){
+        if(projectName == null || projectName.isEmpty()){
+            System.out.println("ID проекта не может быть пустым");
+            return null;
+        }
         return projectRepository.findByName(projectName);
     }
     public Project findById(int projectName){
@@ -58,18 +62,49 @@ public class ProjectService {
     }
 
     //UPDATE
-    public void updateByName(String projectName){
-        //Поле + новое значение
-        projectRepository.updateByName(projectName);
+    public void updateByName(String projectName, String fieldForUpdate, String newValue){
+        if(projectName == null || projectName.isEmpty()){
+            System.out.println("ID проекта не может быть пустым");
+            return;
+        }
+        if(fieldForUpdate == null || fieldForUpdate.isEmpty()){
+            System.out.println("Поле для обновления не может быть пустой");
+            return;
+        }
+        if((newValue == null || newValue.isEmpty()) && !fieldForUpdate.equalsIgnoreCase("описание")){
+            System.out.println("Новое значение не может быть пустым");
+            return;
+        }
+        projectRepository.updateByName(projectName, fieldForUpdate, newValue);
     }
 
-    public void updateById(int projectId){
-        //Поле + новое значение
-        projectRepository.updateById(projectId);
+    public void updateById(String projectId, String fieldForUpdate, String newValue){
+        int projectIdChecked;
+        if(projectId == null || projectId.isEmpty()){
+            System.out.println("ID проекта не может быть пустым");
+            return;
+        }
+        else{
+            projectIdChecked = Integer.parseInt(projectId);
+        }
+        if(fieldForUpdate == null || fieldForUpdate.isEmpty()){
+            System.out.println("Поле для обновления не может быть пустой");
+            return;
+        }
+        if((newValue == null || newValue.isEmpty()) && !fieldForUpdate.equalsIgnoreCase("описание")){
+            System.out.println("Новое значение не может быть пустым");
+            return;
+        }
+
+        projectRepository.updateById(projectIdChecked, fieldForUpdate, newValue);
     }
 
     //DELETE
     public void deleteByName(String projectName){
+        if(projectName == null || projectName.isEmpty()){
+            System.out.println("Имя проекта не может быть пустым");
+            return;
+        }
         Project project = projectRepository.findByName(projectName);
         for(Task task : project.getTasks()){
             taskService.deleteByName(task.getName());
